+++ pkg/goship/server/server_chi.go
@@ -1,4 +1,4 @@
-package goship
+package server
 
 import (
 	"context"
